pyproject.toml
[project]
name = "keihou-bot"
version = "0.1.0"
description = "JMA alert Discord bot for Tokyo's 23 wards"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "tsukuba-denden" }]
keywords = ["JMA", "alerts", "discord", "tokyo", "apscheduler"]

# Runtime dependencies
dependencies = [
  "requests>=2.32.0",
  "discord.py>=2.4.0",
  "lxml>=5.2.1",
  "APScheduler>=3.10.4",
]

[project.urls]
Homepage = "https://github.com/tsukuba-denden/keihou-bot"

# Dev dependencies (uv supports dependency-groups)
[dependency-groups]
dev = [
  "black>=24.0.0",
  "pylint>=3.2.0",
  "pytest>=7.0.0",
  "pytest-asyncio>=0.23.0",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

# You can run with uv:
#   uv sync --all-extras --group dev
